[Fix #4080] Modularizing open api (#4083)
* [Fix #4080] Modularizing open api


* [Fix #4080] Refining dependencies of token-exhange module

---------
