+++ quarkus/addons/common/sonataflow-deployment/src/main/java/org/kie/kogito/quarkus/serverless/workflow/WorkflowHandlerGeneratedFile.java
