+++ quarkus/addons/openapi/runtime/src/main/java/org/kie/kogito/serverless/workflow/openapi/RuntimeConfigCredentialsProvider.java
