+++ quarkus/addons/token-exchange/deployment/pom.xml
@@ -39,18 +39,9 @@
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-arc-deployment</artifactId>
     </dependency>
-    <dependency>
-      <groupId>io.quarkus</groupId>
-      <artifactId>quarkus-oidc-client-deployment</artifactId>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
+     <dependency>
       <groupId>org.apache.kie.sonataflow</groupId>
-      <artifactId>sonataflow-quarkus-deployment</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.quarkiverse.openapi.generator</groupId>
-      <artifactId>quarkus-openapi-generator-deployment</artifactId>
+      <artifactId>sonataflow-addons-quarkus-openapi-deployment</artifactId>
     </dependency>
     <dependency>
       <groupId>org.kie</groupId>
