+++ quarkus/extensions/kogito-quarkus-serverless-workflow-extension/kogito-quarkus-serverless-workflow-deployment/src/main/java/org/kie/kogito/serverless/workflow/parser/schema/ServerlessWorkflowOASFilter.java
@@ -16,14 +16,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.quarkus.serverless.workflow.openapi;
+package org.kie.kogito.serverless.workflow.parser.schema;
 
 import java.util.Map;
 
 import org.eclipse.microprofile.openapi.OASFilter;
 import org.eclipse.microprofile.openapi.models.OpenAPI;
 import org.eclipse.microprofile.openapi.models.media.Schema;
-import org.kie.kogito.serverless.workflow.parser.schema.OpenApiModelSchemaGenerator;
 
 public final class ServerlessWorkflowOASFilter implements OASFilter {
